requirements.txt
Django==6.0.4
djangorestframework==3.17.1
psycopg2-binary==2.9.12
python-dotenv==1.0.0
